+++ win/gdi_hrgn.go
@@ -37,7 +37,10 @@ var _CreateRectRgnIndirect = dll.Gdi32.NewProc("CreateRectRgnIndirect")
 // [CombineRgn]: https://learn.microsoft.com/en-us/windows/win32/api/wingdi/nf-wingdi-combinergn
 func (hRgn HRGN) CombineRgn(src1, src2 HRGN, mode co.RGN) (co.REGION, error) {
 	ret, _, _ := syscall.SyscallN(_CombineRgn.Addr(),
-		uintptr(hRgn), uintptr(src1), uintptr(src2), uintptr(mode))
+		uintptr(hRgn),
+		uintptr(src1),
+		uintptr(src2),
+		uintptr(mode))
 	if ret == 0 {
 		return co.REGION(0), co.ERROR_INVALID_PARAMETER
 	}
@@ -58,7 +61,8 @@ func (hRgn HRGN) DeleteObject() error {
 // [EqualRgn]: https://learn.microsoft.com/en-us/windows/win32/api/wingdi/nf-wingdi-equalrgn
 func (hRgn HRGN) EqualRgn(other HRGN) bool {
 	ret, _, _ := syscall.SyscallN(_EqualRgn.Addr(),
-		uintptr(hRgn), uintptr(other))
+		uintptr(hRgn),
+		uintptr(other))
 	return ret != 0
 }
 
@@ -70,7 +74,8 @@ var _EqualRgn = dll.Gdi32.NewProc("EqualRgn")
 func (hRgn HRGN) GetRgnBox() (RECT, co.REGION, error) {
 	var rc RECT
 	ret, _, _ := syscall.SyscallN(_GetRgnBox.Addr(),
-		uintptr(hRgn), uintptr(unsafe.Pointer(&rc)))
+		uintptr(hRgn),
+		uintptr(unsafe.Pointer(&rc)))
 	if ret == utl.REGION_ERROR {
 		return RECT{}, co.REGION(0), co.ERROR_INVALID_PARAMETER
 	}
@@ -84,7 +89,9 @@ var _GetRgnBox = dll.Gdi32.NewProc("GetRgnBox")
 // [OffsetClipRgn]: https://learn.microsoft.com/en-us/windows/win32/api/wingdi/nf-wingdi-offsetcliprgn
 func (hRgn HRGN) OffsetClipRgn(x, y int32) (co.REGION, error) {
 	ret, _, _ := syscall.SyscallN(_OffsetClipRgn.Addr(),
-		uintptr(hRgn), uintptr(x), uintptr(y))
+		uintptr(hRgn),
+		uintptr(int32(x)),
+		uintptr(int32(y)))
 	if ret == 0 {
 		return co.REGION(0), co.ERROR_INVALID_PARAMETER
 	}
@@ -98,7 +105,9 @@ var _OffsetClipRgn = dll.Gdi32.NewProc("OffsetClipRgn")
 // [OffsetRgn]: https://learn.microsoft.com/en-us/windows/win32/api/wingdi/nf-wingdi-offsetrgn
 func (hRgn HRGN) OffsetRgn(x, y int32) (co.REGION, error) {
 	ret, _, _ := syscall.SyscallN(_OffsetRgn.Addr(),
-		uintptr(hRgn), uintptr(x), uintptr(y))
+		uintptr(hRgn),
+		uintptr(int32(x)),
+		uintptr(int32(y)))
 	if ret == 0 {
 		return co.REGION(0), co.ERROR_INVALID_PARAMETER
 	}
@@ -112,7 +121,9 @@ var _OffsetRgn = dll.Gdi32.NewProc("OffsetRgn")
 // [PtInRegion]: https://learn.microsoft.com/en-us/windows/win32/api/wingdi/nf-wingdi-ptinregion
 func (hRgn HRGN) PtInRegion(x, y int32) bool {
 	ret, _, _ := syscall.SyscallN(_PtInRegion.Addr(),
-		uintptr(hRgn), uintptr(x), uintptr(y))
+		uintptr(hRgn),
+		uintptr(int32(x)),
+		uintptr(int32(y)))
 	return ret != 0
 }
 
@@ -123,7 +134,8 @@ var _PtInRegion = dll.Gdi32.NewProc("PtInRegion")
 // [RectInRegion]: https://learn.microsoft.com/en-us/windows/win32/api/wingdi/nf-wingdi-rectinregion
 func (hRgn HRGN) RectInRegion(rc RECT) bool {
 	ret, _, _ := syscall.SyscallN(_RectInRegion.Addr(),
-		uintptr(hRgn), uintptr(unsafe.Pointer(&rc)))
+		uintptr(hRgn),
+		uintptr(unsafe.Pointer(&rc)))
 	return ret != 0
 }
